requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.30.0


# Conservative Django
django==1.11.9 # pyup: <2.0

# Configuration
django-environ==0.4.4
whitenoise==3.3.1


# Forms
django-crispy-forms==1.7.0

# Models
django-model-utils==3.1.1

# Images
Pillow==5.0.0

# Password storage
argon2-cffi==18.1.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.34.0

# On Windows, you must download/install psycopg2 manually
# from http://www.lfd.uci.edu/~gohlke/pythonlibs/#psycopg


# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2017.3

# Redis support
django-redis==4.8.0
redis==2.10.6


celery==3.1.25



rcssmin==1.0.6 --install-option="--without-c-extensions"
django-compressor==2.2


# Your custom requirements go here
cookiecutter==1.6.0
coverage==4.4.2
decorator==4.2.1
defusedxml==0.5.0
django-appconf==1.0.2
django-bootstrap-datepicker==1.2.3
django-bower==5.2.0
django-coverage-plugin==1.5.0
django-debug-toolbar==1.9.1
django-extensions==1.9.9
django-test-plus==1.0.22
docutils==0.14
factory-boy==2.9.2
Faker==0.8.10
future==0.16.0
icalendar==4.0.0
idna==2.6
imagesize==0.7.1
invoke==0.22.0
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
Jinja2==2.10
jinja2-time==0.2.0
kombu==3.0.37
MarkupSafe==1.0
oauthlib==2.0.6
olefile==0.44
parso==0.1.1
pickleshare==0.7.4
pluggy==0.6.0
poyo==0.4.1
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
py==1.5.2
pycparser==2.18
Pygments==2.2.0
pytest==3.3.2
pytest-django==3.1.2
pytest-sugar==0.9.0
python-dateutil==2.6.1
python3-openid==3.1.0
regex==2018.1.10
requests==2.18.4
requests-oauthlib==0.8.0
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
termcolor==1.1.0
text-unidecode==1.1
traitlets==4.3.2
typing==3.6.2
Unidecode==0.4.21
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
whichcraft==0.4.1
wincertstore==0.2
